Find cliques

Find cliques

clique(A, adjacency_list = FALSE, min = NULL, max = NULL)

Arguments

A

A symmetric matrix object.

adjacency_list

Whether to return the adjacency list of cliques per node.

min

Numeric constant, lower limit on the size of the cliques to find. NULL means no limit, ie. it is the same as 0.

max

Numeric constant, upper limit on the size of the cliques to find. NULL means no limit.

Value

This function return the list of cliques that each node belong. If adjacency_list = TRUE it also return the adjacency list of cliques per node.

References

Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge University Press.

Author

Alejandro Espinosa-Rada

Examples

A <- matrix(c(0,1,1,1,0,
            1,0,1,0,0,
            1,1,0,0,0,
            1,0,0,0,1,
            0,0,0,1,0), byrow = TRUE, ncol = 5)
rownames(A) <- letters[1:nrow(A)]
colnames(A) <- letters[1:ncol(A)]

clique(A, adjacency_list = TRUE, min = 3)
#> $nodes
#>      node CliqueID
#> [1,] "a"  "1"     
#> [2,] "b"  "1"     
#> [3,] "c"  "1"     
#> 
#> $adjacency_list
#> $adjacency_list$a
#> [1] "abc" "abd" "acd"
#> 
#> $adjacency_list$b
#> [1] "abc"
#> 
#> $adjacency_list$c
#> [1] "abc"
#> 
#>