Triad census analysis assuming U|MAN

Considering the triad census of Davis and Leinhardt (1972) for vector A, B, and C:

Usage

triad_uman(A, ztest = FALSE, covar = FALSE)

Arguments

A

A symmetric matrix object

ztest

Return Z and p-value

covar

Return the covariance matrix for triadic analysis

Value

This function gives the counts of the triad census, the expected counts, assuming that U|MAN distribution (Holland and Leinhardt, 1975, 1976) is operating, and the standard deviations of these counts.

Details

003 = A,B,C, empty triad

012 = A -> B, C, triad with a single directed edge

102 = A <-> B, C, triad with a reciprocated connection between two vertices

021D = A <-B-> C, triadic out-star

021U = A -> B <- C triadic in-star

021C = A-> B-> C, directed line

111D = A <-> B <-C

111U = A <-> B-> C

030T = A-> B <-C, A-> C

030C = A <-B <-C, A-> C

201 = A <-> B <-> C

120D = A <-B-> C, A <-> C

120U = A-> B <-C, A <->C

120C = A-> B-> C, A <-> C

210 = A-> B <-> C, A <-> C

300 = A <-> B <-> C, A <->C, complete triad.

References

Davis, J.A. and Leinhardt, S. (1972). The Structure of Positive Interpersonal Relations in Small Groups. In J. Berger (Ed.), Sociological Theories in Progress, Volume 2, 218-251. Boston: Houghton Mifflin.

Holland, P. W. and Leinhardt, S. (1975). The statistical analysis of local structure in social networks. In D. R. Heise (Ed.), Sociological Methodology, 1976 (Jossey-Bass, pp. 1–45).

Holland, P. W. and Leinhardt, S. (1976). Local Structure in Social Networks. Sociological Methodology, 7, 1–45. doi:10.2307/270703

Wasserman, S. and Faust, K. (1994). Social network analysis: Methods and applications. Cambridge University Press.

Author

Alejandro Espinosa-Rada

Examples

data(krackhardt_friends)
triad_uman(krackhardt_friends)
#>    label OBS     EXP     VAR    STD
#> 1    003 376 320.057  88.239  9.394
#> 2    012 366 416.818 211.856 14.555
#> 3    102 143 171.193  88.921  9.430
#> 4   021D 114  44.087  38.736  6.224
#> 5   021U  34  44.087  38.736  6.224
#> 6   021C  35  88.173  66.770  8.171
#> 7   111D  39  73.745  60.470  7.776
#> 8   111U 101  73.745  60.470  7.776
#> 9   030T  23  18.173  14.902  3.860
#> 10  030C   0   6.058   5.694  2.386
#> 11   201  20  28.971  20.471  4.524
#> 12  120D  16   7.740   7.359  2.713
#> 13  120U  25   7.740   7.359  2.713
#> 14  120C   9  15.481  13.957  3.736
#> 15   210  23  12.385  10.534  3.246
#> 16   300   6   1.548   1.441  1.200
# \donttest{
triad_uman(krackhardt_friends, ztest = TRUE, covar = TRUE)
#> $results
#>    label OBS     EXP     VAR    STD      Z     P
#> 1    003 376 320.057  88.239  9.394  5.955 0.000
#> 2    012 366 416.818 211.856 14.555 -3.491 0.000
#> 3    102 143 171.193  88.921  9.430 -2.990 0.003
#> 4   021D 114  44.087  38.736  6.224 11.233 0.000
#> 5   021U  34  44.087  38.736  6.224 -1.621 0.105
#> 6   021C  35  88.173  66.770  8.171 -6.508 0.000
#> 7   111D  39  73.745  60.470  7.776 -4.468 0.000
#> 8   111U 101  73.745  60.470  7.776  3.505 0.000
#> 9   030T  23  18.173  14.902  3.860  1.251 0.211
#> 10  030C   0   6.058   5.694  2.386 -2.539 0.011
#> 11   201  20  28.971  20.471  4.524 -1.983 0.047
#> 12  120D  16   7.740   7.359  2.713  3.045 0.002
#> 13  120U  25   7.740   7.359  2.713  6.362 0.000
#> 14  120C   9  15.481  13.957  3.736 -1.735 0.083
#> 15   210  23  12.385  10.534  3.246  3.270 0.001
#> 16   300   6   1.548   1.441  1.200  3.710 0.000
#> 
#> $z_test
#> z p 
#> 0 1 
#> 
#> $covariance
#>           003     012     102    021D    021U   021C    111D   111U   030T
#> 003    88.239   0.000   0.000   0.000   0.000  0.000   0.000  0.000  0.000
#> 012  -107.176 211.856   0.000   0.000   0.000  0.000   0.000  0.000  0.000
#> 102   -44.019  -3.238  88.921   0.000   0.000  0.000   0.000  0.000  0.000
#> 021D    4.708 -21.425   4.554  38.736   0.000  0.000   0.000  0.000  0.000
#> 021U    4.708 -21.425   4.554  -5.351  38.736  0.000   0.000  0.000  0.000
#> 021C    9.415 -42.851   9.107 -10.701 -10.701 66.770   0.000  0.000  0.000
#> 111D    7.874  -8.646 -19.575  -2.089  -2.089 -4.178  60.470  0.000  0.000
#> 111U    7.874  -8.646 -19.575  -2.089  -2.089 -4.178 -13.275 60.470  0.000
#> 030T    6.647  -5.682   3.555  -3.570  -3.570 -7.140  -1.005 -1.005 14.902
#> 030C    2.216  -1.894   1.185  -1.190  -1.190 -2.380  -0.335 -0.335 -1.090
#> 201     3.094   7.286   7.064   1.115   1.115  2.230  -6.019 -6.019  0.602
#> 120D    2.831  -0.384  -0.521  -0.677  -0.677 -1.355  -1.770 -1.770 -0.627
#> 120U    2.831  -0.384  -0.521  -0.677  -0.677 -1.355  -1.770 -1.770 -0.627
#> 120C    5.662  -0.769  -1.043  -1.355  -1.355 -2.710  -3.539 -3.539 -1.254
#> 210     4.530   2.642  -4.091  -0.065  -0.065 -0.130  -3.658 -3.658 -0.169
#> 300     0.566   0.737  -0.919   0.078   0.078  0.156  -0.396 -0.396  0.032
#>        030C    201   120D   120U   120C    210   300
#> 003   0.000  0.000  0.000  0.000  0.000  0.000 0.000
#> 012   0.000  0.000  0.000  0.000  0.000  0.000 0.000
#> 102   0.000  0.000  0.000  0.000  0.000  0.000 0.000
#> 021D  0.000  0.000  0.000  0.000  0.000  0.000 0.000
#> 021U  0.000  0.000  0.000  0.000  0.000  0.000 0.000
#> 021C  0.000  0.000  0.000  0.000  0.000  0.000 0.000
#> 111D  0.000  0.000  0.000  0.000  0.000  0.000 0.000
#> 111U  0.000  0.000  0.000  0.000  0.000  0.000 0.000
#> 030T  0.000  0.000  0.000  0.000  0.000  0.000 0.000
#> 030C  5.694  0.000  0.000  0.000  0.000  0.000 0.000
#> 201   0.201 20.471  0.000  0.000  0.000  0.000 0.000
#> 120D -0.209 -0.433  7.359  0.000  0.000  0.000 0.000
#> 120U -0.209 -0.433 -0.381  7.359  0.000  0.000 0.000
#> 120C -0.418 -0.865 -0.762 -0.762 13.957  0.000 0.000
#> 210  -0.056 -3.114 -0.583 -0.583 -1.166 10.534 0.000
#> 300   0.011 -0.857 -0.042 -0.042 -0.083 -0.366 1.441
#> 
# }